🔖 Quark v0.2.0
QuarkColorConverter Still w*rking on Quark, the alpha values for the RGB just really isn't needed, so I removed that, and I've added a function so that it doesn't has to be a float number.

Also what was the point of the A value lmao, it's always hardcoded to be 1.0. it only point of converting RGBA to 'Quarks float numbers' was because it's float numbers, it doesn't make sense to convert RGA values to RGA, that's just dumb and doesn't make sense — so right now, it just converts HEX colors to RGA. I'm planning to add more later in the future and like, if you wanna make a PR you can do so lmao-- it's just that you guys can't exec the application because the source is still closed. HOLY AM I YAPPIN-